requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
openai==1.3.7
requests==2.31.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
